+++ stroom-security-server/src/main/java/stroom/security/server/UserAppPermissionServiceImpl.java
@@ -34,7 +34,6 @@ import stroom.util.spring.StroomBeanStore;
 import stroom.util.spring.StroomStartup;
 
 import javax.inject.Inject;
-import javax.persistence.PersistenceException;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -53,6 +52,8 @@ public class UserAppPermissionServiceImpl implements UserAppPermissionService {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserAppPermissionServiceImpl.class);
 
     static {
+        // We are expecting duplicate keys at times so have to protect against it
+        // by only inserting if the compound key doesn't exist already
         SQL_INSERT_USER_PERMISSIONS = ""
                 + "INSERT INTO "
                 + AppPermission.TABLE_NAME
@@ -63,7 +64,19 @@ public class UserAppPermissionServiceImpl implements UserAppPermissionService {
                 + " ,"
                 + AppPermission.PERMISSION
                 + ")"
-                + " VALUES (?,?,?)";
+                + " SELECT ?, ?, ?"
+                + " FROM DUAL"
+                + " WHERE NOT EXISTS ("
+                + " SELECT NULL"
+                + " FROM "
+                + AppPermission.TABLE_NAME
+                + " WHERE "
+                + "  "
+                + AppPermission.USER_UUID
+                + " = ? AND "
+                + AppPermission.PERMISSION
+                + " = ?"
+                + ")";
     }
 
     static {
@@ -235,12 +248,22 @@ public class UserAppPermissionServiceImpl implements UserAppPermissionService {
             LOGGER.error("Unknown permission: " + permission);
         } else {
             try {
-                final SqlBuilder sqlBuilder = new SqlBuilder(SQL_INSERT_USER_PERMISSIONS, 1, userRef.getUuid(), permissionId);
-                entityManager.executeNativeUpdate(sqlBuilder);
-            } catch (final PersistenceException e) {
-                // Expected exception.
-                LOGGER.debug("addPermission()", e);
-                throw e;
+                // bit grim having to repeat the key fields but they are used twice in the sql
+                final SqlBuilder sqlBuilder = new SqlBuilder(
+                        SQL_INSERT_USER_PERMISSIONS,
+                        1,
+                        userRef.getUuid(), permissionId,
+                        userRef.getUuid(), permissionId);
+
+                final long insertCount = entityManager.executeNativeUpdate(sqlBuilder);
+
+                if (insertCount == 0) {
+                    LOGGER.debug("Key already exists for userRef [{}], permission [{}], nothing inserted",
+                            userRef.getUuid(), permission);
+                } else {
+                    LOGGER.debug("Inserted {} row with userRef [{}], permission [{}]",
+                            insertCount, userRef.getUuid(), permission);
+                }
             } catch (final RuntimeException e) {
                 LOGGER.error("addPermission()", e);
                 throw e;